Add config.py and rename do_meili_tools to utils
